requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
djangorestframework~=3.14.0
python-dotenv~=1.0.0
django-rest-swagger~=2.2.0
drf-yasg~=1.21.7
djangorestframework-simplejwt~=5.3.1
django-debug-toolbar~=4.2.0